Add files via upload
This is a BMI calculator app, which calculates the BMI of the user, and outputs a small text based on the results of the calculation. With beautiful matching colors and a great UI, i learnt a lot of flutter concepts from this project like; using plug-ins in your app, refactoring flutter widgets, the GestureDetector widget, and a great deal of others.